Move "is_current_alert" function into model
This will make it easier to ask each alert is it's current, which
we'll need to know when adding the previous alerts feature.
